add calculated data descriptions, rename adapter files
