+++ tests/parsers/unchained_labs_lunatic_stunner/unchained_labs_lunatic_structure_test.py
@@ -15,13 +15,13 @@ from allotropy.parsers.unchained_labs_lunatic_stunner.constants import (
     NO_DATE_OR_TIME_ERROR_MSG,
     NO_MEASUREMENT_IN_PLATE_ERROR_MSG,
 )
-from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_calcdocs import (
+from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_stunner_calcdocs import (
     create_calculated_data,
 )
-from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_reader import (
+from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_stunner_reader import (
     UnchainedLabsLunaticReader,
 )
-from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_structure import (
+from allotropy.parsers.unchained_labs_lunatic_stunner.unchained_labs_lunatic_stunner_structure import (
     _create_measurement,
     _create_measurement_group,
     create_measurement_groups,
